PUBLIC: manually install the correct version of numpy
The second to last cell of the colab updates numpy to version 2.0 which is incompatible with the API used in the second cell of the colab. Therefore, when we run the colab for the second time, the colab will yield an ImportError and an AttributeError as shown in b/349459984 and cannot proceed anymore.